Fixed Circular Dependency to Access AuthService
