+++ apps/koala-frontend/src/app/core/components/header/header.component.ts
@@ -7,8 +7,8 @@ import { AuthService } from '../../../features/auth/services/auth.service';
 import { environment } from '../../../../environments/environment';
 import { NavigationService } from '../../../features/sessions/services/navigation.service';
 import { SessionsService } from '../../../features/sessions/services/sessions.service';
-import { ApolloQueryResult } from '@apollo/client/core';
-import { GetOneSessionQuery } from '../../../graphql/generated/graphql';
+import { Session } from '../../../features/sessions/types/session.entity';
+import { AccessTokenService } from '../../../features/auth/services/access-token.service';
 
 export enum LANGUAGE_CODE {
   GERMAN = 'de',
@@ -51,7 +51,8 @@ export class HeaderComponent {
     private readonly translateService: TranslateService,
     private readonly authService: AuthService,
     private readonly navigationService: NavigationService,
-    private readonly sessionService: SessionsService
+    private readonly sessionService: SessionsService,
+    private readonly accessTokenService: AccessTokenService
   ) {
     this.router.events.pipe(filter((event: any) => event instanceof NavigationEnd)).subscribe((event) => {
       const routeUrl: string = event.url;
@@ -61,10 +62,9 @@ export class HeaderComponent {
         if (sessionIdInURL) {
           this.sessionId = parseInt(sessionIdInURL[0]);
           this.sessionService.getOne(this.sessionId).subscribe({
-            next: (response: ApolloQueryResult<GetOneSessionQuery>) => {
-              const selectedSession = response.data.session;
-
-              this.isUserSessionOwner = selectedSession.owner.id === this.authService.getLoggedInUserId().toString();
+            next: (selectedSession: Session) => {
+              this.isUserSessionOwner =
+                selectedSession.owner?.id === this.accessTokenService.getLoggedInUserId().toString();
             },
           });
         }
